+++ docs/src/release_notes.rst
@@ -27,6 +27,9 @@ Releases of documentation and code are using the same version numbers.
 
   * New: Added optional simulation timeout for the :ref:`run(until_change)
     <sec_tcp_cmd_run>` command.
+  * New: Added new supported behavior for the :ref:`run(until_change)
+    <sec_tcp_cmd_run>` command; it can now also set the simulation to run until the
+    variable changes to any value.
   * New: Added the :json:`"type": "timeout"` :ref:`return message
     <sec_tcp_ret_msg>` type.
   * New: Added :json:`"verisocks_version"` field in the return message for the
@@ -44,6 +47,8 @@ Releases of documentation and code are using the same version numbers.
     <verisocks.verisocks.Verisocks.set_value>`
   * New: Added shortcut function :py:func:`Verisocks.get_value()
     <verisocks.verisocks.Verisocks.get_value>`
+  * New: Supported change to the :ref:`run(until_change) <sec_tcp_cmd_run>`
+    command.
 
 1.6.0 - 2026-07-26
 ******************